\name{dic}
\alias{dic}
\title{
English to Persian dictionary
}
\description{
Translate words from English to Persian (Over 67,000 words)
}
\usage{
dic("x")
}
\arguments{
  \item{x}{
    any english word
}
}
\value{
An object of class \code{dic}, if possible, or an error message if
  not.
}
\references{
SSee original note by aryanpour at
\url{http://www.aryanpour.com/}.
}
\author{
Sadegh Rasoulinejad ,MSc Student of Biostatistics 
  Mashhad University of Medical Sciences (MUMS) (Using dictionaries aryanpour)
}

\examples{

dic("help")

}
\keyword{ package }
