\name{persiandictionary-package}
\alias{persiandictionary-package}
\alias{persiandictionary}
\docType{package}
\title{
persiandictionary
}
\description{
Translate words from English to Persian (Over 67,000 words)
}
\details{
  \tabular{ll}{
   Package: \tab dictionary\cr
   Type: \tab Package\cr
   Version: \tab 1.0\cr
   Date: \tab 2013-03-28\cr
   License: \tab GPL-2\cr
   LazyLoad: \tab yes\cr
}
In answer to a question on R-help Sadegh Rasoulinejad provided an elegant R
function to Translate words from English to Persian.
}
\author{
Sadegh Rasoulinejad (Using dictionaries aryanpour)
  MSc Student of Biostatistics 
  Mashhad University of Medical Sciences (MUMS)

Maintainer: Sadegh Rasoulinejad, <RasouliNS911@mums.ac.ir >

}
\references{
See original note by aryanpour at
\url{http://www.aryanpour.com/}.
}

\keyword{ package }
\seealso{
\code{\link[utils]{as.roman}}.
}
\examples{
 dic("help")
}
